Move SingleCertVerifier into its own module
